Added unit tests for util functions separately
