+++ packages/jaeger-ui/src/components/TracePage/CriticalPath/utils/findLastFinishingChildSpanId.tsx
@@ -1,0 +1,35 @@
+// Copyright (c) 2023 Uber Technologies, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import { Span, Trace } from '../../../../types/trace';
+
+// This function finds LFC of a current span which is less than the spawn event time
+// |-------------spanA--------------|
+//    |--spanB--|    |--spanC--|
+// Here span-C will be LFC of spanA. But spanA doesnt have LFC.So, it returns to parent
+// from its creationTime (startTime) and this is referred as spawnTime below
+const findLastFinishingChildSpanId = (
+  traceData: Trace,
+  currentSpan: Span,
+  spawnTime?: number
+): string | undefined => {
+  if (spawnTime) {
+    return currentSpan.childSpanIds.find(each =>
+      traceData.spans.some(span => span.spanID === each && span.startTime + span.duration < spawnTime)
+    );
+  }
+  return currentSpan.childSpanIds ? currentSpan.childSpanIds[0] : undefined;
+};
+
+export default findLastFinishingChildSpanId;
