+++ packages/jaeger-ui/src/components/TracePage/CriticalPath/utils/findRootSpanId.tsx
@@ -1,0 +1,22 @@
+// Copyright (c) 2023 Uber Technologies, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import { Span } from '../../../../types/trace';
+
+const findRootSpanId = (spans: Span[]): string | undefined => {
+  const rootSpan = spans.find(span => span.references.length === 0 && span.depth === 0);
+  return rootSpan?.spanID;
+};
+
+export default findRootSpanId;
